requirements.txt
requests
psycopg2-binary
python-dotenv
boto3
polars
duckdb
pyarrow
dbt-duckdb